refactor(ts-bindings): export bindings from a binary in hc
Replace the aggregate export test in holochain_conductor_api with an
`export-ts-bindings` binary in the hc crate, gated on
`required-features = ["ts_rs"]` so no ordinary build produces it. The
generated tree is byte-identical to the one the test wrote.

The binary lives in hc so that a consumer can build it through Holonix,
whose hc package appends overridden arguments to a hard-wired
`--manifest-path crates/hc/Cargo.toml --bin hc`; a binary in any other
crate cannot be named alongside those. hc also pulls in neither the
holochain crate nor a wasmer backend, so the export no longer compiles a
WASM engine in order to emit TypeScript.

Set the TypeScript dialect in code rather than reading it from the
environment. Cargo injects the TS_RS_* entries of .cargo/config.toml only
into processes it spawns from this workspace, so the binary run from
anywhere else would have silently emitted bigint and extensionless
import paths.
