+++ Makefile
@@ -124,16 +124,16 @@ test-workspace-wasmer-wasmi:
 		--no-default-features \
 		--features $(DEFAULT_FEATURES),wasmer-wasmi
 
-# Runs the aggregate export test in holochain_conductor_api, which stages
-# the whole binding tree in one process (ts-rs only merges declarations
-# sharing an output file within a single process) before atomically
-# replacing TS_RS_EXPORT_DIR. See holochain_conductor_api::export_ts_bindings.
+# Runs the `export-ts-bindings` binary in the hc crate, which stages the whole
+# binding tree in one process (ts-rs only merges declarations sharing an output
+# file within a single process) before atomically replacing TS_RS_EXPORT_DIR.
+# See holochain_conductor_api::export_ts_bindings.
 ts-bindings:
 	TS_RS_EXPORT_DIR=$(or $(TS_RS_EXPORT_DIR),./bindings) \
-		cargo test -p holochain_conductor_api --features ts_rs,unstable-countersigning export_bindings_aggregate
+		cargo run -p holochain_cli --features ts_rs,unstable-countersigning --bin export-ts-bindings
 
-# ts_rs-gated tests excluded from the `export_bindings_aggregate` name filter
-# above: the wire-format smoke test and the tag-injection helper's unit tests.
+# The ts_rs-gated tests in holochain_conductor_api: the wire-format smoke test
+# and the tag-injection helper's unit tests.
 ts-bindings-test: ts-bindings
 	cargo test -p holochain_conductor_api --features ts_rs,unstable-countersigning admin_request_bindings_smoke
 	cargo test -p holochain_conductor_api --features ts_rs,unstable-countersigning tag_exports_public
