+++ crates/hc/Cargo.toml
@@ -18,6 +18,13 @@ description = "Provides the `hc` binary, a helpful CLI tool for working with Hol
 name = "hc"
 path = "src/bin/hc.rs"
 
+# Writes the TypeScript binding tree. Skipped by builds that don't enable
+# `ts_rs`, including `cargo build --workspace`.
+[[bin]]
+name = "export-ts-bindings"
+path = "src/bin/export-ts-bindings.rs"
+required-features = ["ts_rs"]
+
 [lib]
 name = "holochain_cli"
 path = "src/lib.rs"
@@ -32,6 +39,10 @@ holochain_cli_sandbox = { path = "../hc_sandbox", version = "^0.8.0-dev.1", defa
 holochain_cli_client = { path = "../hc_client", version = "^0.8.0-dev.1", default-features = false }
 holochain_trace = { version = "^0.8.0-dev.0", path = "../holochain_trace" }
 tokio = { version = "1.36.0", features = ["full"] }
+
+# Optional, for the `export-ts-bindings` binary only
+holochain_conductor_api = { version = "^0.8.0-dev.1", path = "../holochain_conductor_api", default-features = false, optional = true }
+ts-rs = { version = "12.0", optional = true }
 
 [dev-dependencies]
 assert_cmd = "2.2"
@@ -52,3 +63,14 @@ wasmer-wasmi = [
   "holochain_cli_sandbox/wasmer-wasmi",
   "holochain_cli_client/wasmer-wasmi",
 ]
+
+# Builds the `export-ts-bindings` binary, which writes the TypeScript
+# bindings for the conductor's wire API.
+ts_rs = [
+  "dep:holochain_conductor_api",
+  "dep:ts-rs",
+  "holochain_conductor_api/ts_rs",
+]
+
+# Includes the countersigning app API in the exported TypeScript bindings.
+unstable-countersigning = ["holochain_conductor_api?/unstable-countersigning"]
